MainActivity layout added. Androidx dependencies added for recyclerview and cardview.
