--- app/src/main/java/org/ohmstheresistance/mastermindmvp/activities/MainActivity.java
@@ -1,16 +1,0 @@
-package org.ohmstheresistance.mastermindmvp.activities;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-import org.ohmstheresistance.mastermindmvp.R;
-
-public class MainActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.main_activity);
-    }
-}
