+++ app/src/main/java/org/ohmstheresistance/mastermindmvp/activities/activities/MainActivity.java
@@ -1,4 +1,4 @@
-package org.ohmstheresistance.mastermindmvp.activities;
+package org.ohmstheresistance.mastermindmvp.activities.activities;
 
 import androidx.appcompat.app.AppCompatActivity;
 
@@ -6,11 +6,11 @@ import android.os.Bundle;
 
 import org.ohmstheresistance.mastermindmvp.R;
 
-public class MainPageActivity extends AppCompatActivity {
+public class MainActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.main_page_activity);
+        setContentView(R.layout.main_activity);
     }
 }
