+++ app/src/main/java/org/ohmstheresistance/mastermindmvp/activities/contract/MainContract.java
@@ -1,0 +1,12 @@
+package org.ohmstheresistance.mastermindmvp.activities.contract;
+
+public interface MainContract {
+
+    interface MastermindView {
+
+    }
+
+    interface MastermindPresenter{
+
+    }
+}
