+++ app/src/main/res/layout/main_page_activity.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryDark"
-    tools:context=".activities.MainPageActivity">
+    tools:context=".activities.activities.MainPageActivity">
 
 
     <TextView
